backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
google-auth-oauthlib
google-api-python-client
